Reorganize classes into appropriate directories
