+++ PermissionsScraper/PermissionsScraper/Services/AuthService.cs
@@ -1,0 +1,28 @@
+﻿using Microsoft.Identity.Client;
+using System;
+using PermissionsScraper.Common;
+
+namespace PermissionsScraper.Services
+{
+    public static class AuthService
+    {
+        /// <summary>
+        /// Gets authentication to a protected Web API.
+        /// </summary>
+        /// <param name="config">The application configuration settings.</param>
+        /// <returns>An authentication result, if successful.</returns>
+        public static AuthenticationResult GetAuthentication(ApplicationConfig config)
+        {
+            IConfidentialClientApplication app;
+            app = ConfidentialClientApplicationBuilder.Create(config.ClientId)
+                  .WithClientSecret(config.ClientSecret)
+                  .WithAuthority(new Uri(config.Authority))
+                  .Build();
+
+            string[] scopes = new string[] { $"{config.ApiUrl}.default" };
+
+            return app.AcquireTokenForClient(scopes)
+              .ExecuteAsync().GetAwaiter().GetResult();
+        }
+    }
+}
